refactor code to llm and challenges and update readme
